Created check to see if we do not have enough GPU memory. Works but need to be adapted so eacvh thread has its onw part_buffer_size

+++ src/gpu_pack_params.h
@@ -41,7 +41,7 @@ struct gpu_global_pack_params {
   int n_bundles;
 
   /*! Size of the particle data buffers for offloading */
-  int part_buffer_size;
+  long part_buffer_size;
 
   /*! Size of the particle data buffers for offloading */
   int part_send_size_d;
